+++ Application/index.html
@@ -10,5 +10,5 @@
 <body>
     <div class="wrapper"></div>
 </body>
-<script src="nothing2.js" type="module"></script>
+<script src="main.js" type="module"></script>
 </html>
